Added `ser/tuples` module
